requirements.txt
# Web Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.3
Werkzeug==3.0.1
WTForms==3.1.1

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
shap==0.43.0

# Utilities
geopy==2.4.1
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0

# Development Tools (optional)
flask-cors==4.0.0
